Update sync script to skip pull&push if the layers are the same from the src to the target. Add skopeo to github actions

+++ single_sync.sh
@@ -11,15 +11,28 @@ do
     for (( t=0; t<${TAG}; t++ ))
     do
         LOCAL_TAG=$(yq e '.images['"${c}"'].tag['"${t}"']' $1)
-        docker pull ${SRC}:${LOCAL_TAG}
-        for (( d=0; d<${DST}; d++ ))
-          do
-            TO=$(yq e '.images['"${c}"'].destinations['"${d}"']' $1):${LOCAL_TAG}
-            docker tag ${SRC}:${LOCAL_TAG} ${TO}
-            docker push ${TO}
-            docker rmi ${TO}
-          done
-        docker rmi ${SRC}:${LOCAL_TAG}
+        LOCAL_LAYERS=$(skopeo inspect --override-os linux docker://${SRC}:${LOCAL_TAG} 2> /dev/null | yq .Layers)
+        TARGET_LAYERS=$(skopeo inspect --override-os linux docker://$(yq e '.images['"${c}"'].destinations[0]' $1):${LOCAL_TAG} 2> /dev/null | yq .Layers)
+
+        diff <(echo ${LOCAL_LAYERS}) <(echo ${TARGET_LAYERS}) > /dev/null
+
+        if [ $? -eq 0 ]; then
+          echo "    - Skipping ${SRC}:${LOCAL_TAG} as it is already synced"
+        else
+          echo "    - Layer are different, syncing ${SRC}:${LOCAL_TAG}"
+          docker pull ${SRC}:${LOCAL_TAG}
+          for (( d=0; d<${DST}; d++ ))
+            do
+              TO=$(yq e '.images['"${c}"'].destinations['"${d}"']' $1):${LOCAL_TAG}
+              docker tag ${SRC}:${LOCAL_TAG} ${TO}
+              docker push ${TO}
+              docker rmi ${TO}
+            done
+          docker rmi ${SRC}:${LOCAL_TAG}
+        fi
     done
     echo "  - Finish ${NAME}"
 done
+
+# check if two string are equals
+# if [ "$1" == "$2" ]; then
